Respond to review comments: * Add note about throwing ExecutionException * Move some methods * Rename monitoring_process -> monitoring_subprocess
