+++ mlflow/projects/pollable_run.py
@@ -1,11 +1,8 @@
 from abc import abstractmethod
 import os
-import signal
 import subprocess
-import sys
 
 from mlflow.entities.run_status import RunStatus
-from mlflow.utils.logging_utils import eprint
 
 
 def maybe_set_run_terminated(active_run, status):
@@ -15,31 +12,6 @@ def maybe_set_run_terminated(active_run, status):
     """
     if active_run and not RunStatus.is_terminated(active_run.get_run().info.status):
         active_run.set_terminated(status)
-
-
-def monitor_run(pollable_run, active_run):
-    """
-    Polls the run for termination, sending updates on the run's status to a tracking server via
-    the passed-in `ActiveRun` instance. This function is intended to be run asynchronously
-    in a subprocess.
-    """
-    # Add a SIGTERM & SIGINT handler to the current process that cancels the run
-    def handler(signal_num, stack_frame):  # pylint: disable=unused-argument
-        eprint("=== Run (%s) was interrupted, cancelling run... ===" % pollable_run.describe())
-        pollable_run.cancel()
-        maybe_set_run_terminated(active_run, "FAILED")
-        sys.exit(0)
-    signal.signal(signal.SIGTERM, handler)
-    signal.signal(signal.SIGINT, handler)
-    # Perform any necessary setup for the pollable run, then wait on it to finish
-    pollable_run.setup()
-    run_succeeded = pollable_run.wait()
-    if run_succeeded:
-        eprint("=== Run (%s) succeeded ===" % pollable_run.describe())
-        maybe_set_run_terminated(active_run, "FINISHED")
-    else:
-        eprint("=== Run (%s) failed ===" % pollable_run.describe())
-        maybe_set_run_terminated(active_run, "FAILED")
 
 
 class PollableRun(object):
@@ -82,22 +54,6 @@ class PollableRun(object):
         pass
 
 
-def _launch_command(command, work_dir, env_map, stream_output):
-    """
-    Launch entry point command in a subprocess, returning a `subprocess.Popen` representing the
-    subprocess.
-    """
-    cmd_env = os.environ.copy()
-    cmd_env.update(env_map)
-    if stream_output:
-        return subprocess.Popen([os.environ.get("SHELL", "bash"), "-c", command],
-                                cwd=work_dir, env=cmd_env, preexec_fn=os.setsid)
-    return subprocess.Popen(
-        [os.environ.get("SHELL", "bash"), "-c", command],
-        cwd=work_dir, env=cmd_env, stdout=open(os.devnull, 'wb'),
-        stderr=open(os.devnull, 'wb'), preexec_fn=os.setsid)
-
-
 class LocalPollableRun(PollableRun):
     """
     Instance of PollableRun corresponding to a subprocess launched to run an entry point command
@@ -111,9 +67,23 @@ class LocalPollableRun(PollableRun):
         self.stream_output = stream_output
         self.command_proc = None
 
+    def _launch_command(self):
+        """
+        Launch entry point command in a subprocess, returning a `subprocess.Popen` representing the
+        subprocess.
+        """
+        cmd_env = os.environ.copy()
+        cmd_env.update(self.env_map)
+        if self.stream_output:
+            return subprocess.Popen([os.environ.get("SHELL", "bash"), "-c", self.command],
+                                    cwd=self.work_dir, env=cmd_env, preexec_fn=os.setsid)
+        return subprocess.Popen(
+            [os.environ.get("SHELL", "bash"), "-c", self.command],
+            cwd=self.work_dir, env=cmd_env, stdout=open(os.devnull, 'wb'),
+            stderr=open(os.devnull, 'wb'), preexec_fn=os.setsid)
+
     def setup(self):
-        self.command_proc = _launch_command(
-            self.command, self.work_dir, self.env_map, self.stream_output)
+        self.command_proc = self._launch_command()
 
     def wait(self):
         return self.command_proc.wait() == 0
